Employee hierarchy dropped

+++ src/main/java/com/goit/popov/restaurant/controller/PositionController.java
@@ -5,8 +5,10 @@ import com.goit.popov.restaurant.model.Position;
 import com.goit.popov.restaurant.service.PositionService;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
@@ -23,7 +25,9 @@ public class PositionController {
 
         static Logger logger = (Logger) LoggerFactory.getLogger(PositionController.class);
 
-        private static final String DELETION_FAILURE_MESSAGE = "Failed to delete the position!";
+        private static final String FORBIDDEN_ACTION_MESSAGE = "You cannot perform this operation!";
+        private static final String UNIQUE_CONSTRAINT_MESSAGE = "The value entered is not unique!";
+        private static final String ANY_CONSTRAINT_MESSAGE = "Constraint violation error!";
 
         @Autowired
         private PositionService positionService;
@@ -40,31 +44,58 @@ public class PositionController {
         }
 
         @PostMapping(value="/admin/save_position")
-        public String savePosition(@Valid @ModelAttribute("position") Position position, BindingResult result){
+        public String savePosition(@Valid @ModelAttribute("position") Position position, BindingResult result, Model model){
                 if (result.hasErrors()) {
                         logger.error("Errors number while saving new position: "+
                                 result.getFieldErrorCount());
                         return "th/manager/new_position";
                 }
-                positionService.save(position);
+                try {
+                        positionService.save(position);
+                } catch (DataIntegrityViolationException e) {
+                        logger.error("ERROR: "+e.getMessage()+" cause: "+e.getClass());
+                        model.addAttribute("error", UNIQUE_CONSTRAINT_MESSAGE);
+                        return "th/manager/new_position";
+                } catch (Exception e) {
+                        logger.error("ERROR: "+e.getMessage()+" cause: "+e.getClass());
+                        model.addAttribute("error", FORBIDDEN_ACTION_MESSAGE);
+                        return "th/manager/new_position";
+                }
                 return "redirect:/admin/positions";
         }
 
         @GetMapping(value = "/admin/edit_position/{id}")
-        public String showPositionEditForm(@PathVariable("id") int id, ModelMap map){
+        public String showPositionEditForm(@PathVariable("id") int id, ModelMap map, RedirectAttributes ra){
+                try {
+                        if (!positionService.isPossibleOperation(id))
+                                throw new UnsupportedOperationException("You cannot change privileges of this position");
+                } catch (UnsupportedOperationException e) {
+                        ra.addFlashAttribute("error", e.getMessage());
+                        return "redirect:/error";
+                }
                 Position position = positionService.getById(id);
                 map.addAttribute("position", position);
                 return "th/manager/update_position";
         }
 
         @PostMapping(value="/admin/update_position")
-        public String editSave(@Valid @ModelAttribute("position") Position position, BindingResult result){
+        public String updatePosition(@Valid @ModelAttribute("position") Position position, BindingResult result, Model model){
                 if (result.hasErrors()) {
                         logger.error("Errors number while updating position: "+
                                 result.getFieldErrorCount());
                         return "th/manager/update_position";
                 } else {
-                        positionService.update(position);
+                        try {
+                                positionService.update(position);
+                        } catch (DataIntegrityViolationException e) {
+                                logger.error("ERROR: "+e.getMessage()+" cause: "+e.getClass());
+                                model.addAttribute("error", UNIQUE_CONSTRAINT_MESSAGE);
+                                return "th/manager/update_position";
+                        } catch (Exception e) {
+                                logger.error("ERROR: "+e.getMessage()+" cause: "+e.getClass());
+                                model.addAttribute("error", ANY_CONSTRAINT_MESSAGE);
+                                return "th/manager/update_position";
+                        }
                         return "redirect:/admin/positions";
                 }
         }
@@ -72,12 +103,19 @@ public class PositionController {
         @RequestMapping(value="/admin/delete_position/{id}",method = RequestMethod.GET)
         public String delete(@PathVariable int id, RedirectAttributes ra){
                 try {
+                        if (!positionService.isPossibleOperation(id))
+                                throw new UnsupportedOperationException("You cannot delete this position!");
                         positionService.deleteById(id);
+                } catch (UnsupportedOperationException e) {
+                        ra.addFlashAttribute("error", e.getMessage());
+                        logger.error(FORBIDDEN_ACTION_MESSAGE +e.getMessage()+
+                                "/ exception name is: "+ e.getClass());
+                        return "redirect:/error";
                 } catch (Exception e) {
                         ra.addFlashAttribute("status", HttpStatus.FORBIDDEN);
-                        ra.addFlashAttribute("error", "Constraint violation exception deleting position!");
-                        ra.addFlashAttribute("message", DELETION_FAILURE_MESSAGE +id);
-                        logger.error("Constraint violation exception deleting position: "+e.getMessage()+
+                        ra.addFlashAttribute("error", "This position seemingly has references!");
+                        ra.addFlashAttribute("message", ANY_CONSTRAINT_MESSAGE);
+                        logger.error("ERROR: " +e.getMessage()+
                                 "/ exception name is: "+ e.getClass());
                         return "redirect:/error";
                 }
